test: added isolated tests
